requirements.txt
requests == 2.22.0
beautifulsoup4 == 4.8.1
pandas == 0.25.3
sqlalchemy == 1.3.11
psycopg2-binary